+++ caqtus/types/data/_data_type.py
@@ -1,8 +1,13 @@
 from __future__ import annotations
 
+import abc
 from typing import TypeAlias
 
+import attrs
 import numpy as np
+import polars
+
+from caqtus.types.image.roi import RectangularROI
 
 #: A type alias for numpy arrays.
 type Array = np.ndarray
@@ -29,3 +34,32 @@ Only objects that are instances of this type can be saved for a shot.
 
 Note that it is not possible to have structured data containing arrays.
 """
+
+
+class DataType(abc.ABC):
+    @abc.abstractmethod
+    def to_polars_dtype(self) -> polars.DataType:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def to_polars_value(self, value: Data):
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def is_saved_as_array(self) -> bool:
+        raise NotImplementedError
+
+
+@attrs.frozen
+class ImageType(DataType):
+    inner: polars.DataType
+    roi: RectangularROI
+
+    def to_polars_dtype(self) -> polars.DataType:
+        return polars.Array(self.inner, (self.roi.width, self.roi.height))
+
+    def to_polars_value(self, value: Data):
+        return value
+
+    def is_saved_as_array(self) -> bool:
+        return True