+++ changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.24.0] - 2025-03-15
+
+### Added
+
+- Method `Sequence.scan` to lazily load data into a polars DataFrame
+
+### Deprecated
+
+- Method `Sequence.load_shots_data` in favor of `Sequence.scan`
+
 ## [6.23.2] - 2025-03-01
 
 ### Fixed